Adjust solder stop opening to 0.05 and add solder stencil (Paste) layer gerber generation.
As mentioned in #70 the default KiCad 0.2mm solder stop opening is
excessive for most SMD based PCB designs. This can lead to exposed
ground planes and other issues. Also this prevents the solder stop to
create bridges between QFN pads making reliable reflow difficult. A more
common and accepted opening value is 0.05mm.

For PCB assembly a stencil is a very useful tool. For that it is nice to
have the Paste layer gerbers.