Small changes to LCD font
Distinguishing U, V and W was a pain. Switched W to be the same
as inverted M and distinguished U and V by case.

Also, we have enough pixels for nice uppercase Y and lowercase t.

K and X are only pair that seem a bit confusing, but . is ok I guess.