Replaced all 0201 parts except the 0.1u for 0402.
I did not change the 0.1u yet as we have 3 of them on the back side of
the board. They are decoupling caps so they should be as close as
possible to the power pins. So moving them to a spot that can accept a
higher part might be difficult. Although we could use 0402 that are
<0.33mm thick too. This would fit our budget and make assembly easier.
Thus I will make the change in a separate commit.